+++ src/transcriptome_align_exon_overlap.rs
@@ -26,7 +26,7 @@ pub struct GeneStrOverlapAlign {
 #[doc = "Original `Transcriptome::alignExonOverlap` at STAR/source/Transcriptome_alignExonOverlap.cpp:10. Args: nA: uint, aAll: Transcript, strandType: int32, annFeat: ReadAnnotFeature"]
 pub fn transcriptome_alignexonoverlap_l10_transcriptome_alignexonoverlap(
     transcriptome: &crate::transcriptome::Transcriptome,
-    n_a: u32,
+    n_a: u64,
     a_all: &[crate::transcript::Transcript],
     strand_type: i32,
     ann_feat: &mut crate::read_annotations::ReadAnnotFeature,
